requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
breadability==0.1.20
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
chardet==5.1.0
charset-normalizer==3.1.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
daphne==4.0.0
Django==4.1.5
django-filter==23.1
django-grappelli==3.0.5
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.10.7
fonttools==4.39.3
huggingface-hub==0.13.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
mathparse==0.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
nltk==3.5
numpy==1.24.2
packaging==23.0
Pillow==9.5.0
Pint==0.20.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pymongo==3.13.0
pymystem3==0.2.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
pywin32==306
PyYAML==3.13
regex==2023.3.23
requests==2.28.2
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.2.19
sqlparse==0.4.3
sumy==0.11.0
sympy==1.11.1
textblob==0.17.1
tokenizers==0.13.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.15
zope.interface==5.5.2
